Update create_raid_list_json_and_export_it.yml
Waiting for the first job to be finished is not enough to upload the updated file. Because I still do not know how to commit to two different repositories in the same job (for the moment, I can only commit to the first cheched out one), I end up running create_raid_list.py twice, which is dumb.