Rebase before updating raid-list.json in the main repository
Various workflows can run at the same time. If the first one brings changes to the repository, then the second one cannot push local changes without rebasing, which in turn is not possible without commiting or stashing.
I have chosen to rebase after committing... while automatically stashing, in case not all modified files were committed (in the current case, because this would not be relevant).